refactor(domain): replace anemic models with rich domain entities
- Add static create() factory methods
- Remove model classes in favor of entities with behavior
- Convert Answer, AnswerComment, Comment, QuestionComment, User
