+++ src/domain/entities/answer/ports/answer.props.ts
@@ -1,4 +1,4 @@
-import type { Optional } from '@/util/types/optional'
 import type { Answer } from '../answer.entity'
+import type { Optional } from '@/shared/types/common/optional'
 
 export type AnswerProps = Optional<Omit<Answer, 'id' | 'excerpt'>, 'createdAt' | 'updatedAt'>
