+++ src/domain/entities/comment/ports/comment.props.ts
@@ -1,4 +1,4 @@
-import type { Optional } from '@/util/types/optional'
 import { Comment } from '../comment.entity'
+import type { Optional } from '@/shared/types/common/optional'
 
 export type CommentProps = Optional<Omit<Comment, 'id'>, 'createdAt' | 'updatedAt'>
