+++ src/domain/entities/user/ports/user.props.ts
@@ -1,4 +1,4 @@
-import type { Optional } from '@/util/types/optional'
 import type { User } from '../user.entity'
+import type { Optional } from '@/shared/types/common/optional'
 
 export type UserProps = Optional<Omit<User, 'id'>, 'createdAt' | 'updatedAt'>
