--- src/domain/models/answer-comment/answer-comment.model.ts
@@ -1,12 +1,0 @@
-import { Comment } from '../comment/comment.model'
-
-export class AnswerComment extends Comment {
-  constructor (
-    readonly authorId: string,
-    readonly content: string,
-    readonly answerId: string,
-    id?: string
-  ) {
-    super(authorId, content, id)
-  }
-}
