--- src/domain/models/question-comment/question-comment.model.ts
@@ -1,12 +1,0 @@
-import { Comment } from '../comment/comment.model'
-
-export class QuestionComment extends Comment {
-  constructor (
-    readonly authorId: string,
-    readonly content: string,
-    readonly questionId: string,
-    id?: string
-  ) {
-    super(authorId, content, id)
-  }
-}
