--- src/domain/models/user/user.model.ts
@@ -1,12 +1,0 @@
-import { Entity } from '@/core/domain/entity'
-
-export class User extends Entity {
-  constructor (
-    readonly name: string,
-    readonly email: string,
-    readonly password: string,
-    id?: string
-  ) {
-    super(id)
-  }
-}
